Rename sample.
